+++ CHANGELOG.md
@@ -576,6 +576,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Deprecated `activities` in the endpoint `GET api/v1/portfolio/holding/:dataSource/:symbol`
+- Moved the data service to `@ghostfolio/ui/services`
+- Refactored the dividend import
 
 ## 2.228.0 - 2026-01-03